+++ src/crawlee/basic_crawler/basic_crawler.py
@@ -222,6 +222,11 @@ class BasicCrawler(Generic[TCrawlingContext]):
         self._has_finished_before = False
 
     @property
+    def log(self) -> logging.Logger:
+        """The logger used by the crawler."""
+        return self._logger
+
+    @property
     def router(self) -> Router[TCrawlingContext]:
         """The router used to handle each individual crawling request."""
         if self._router is None:
@@ -374,7 +379,8 @@ class BasicCrawler(Generic[TCrawlingContext]):
         self._has_finished_before = True
 
         final_statistics = self._statistics.calculate()
-        self._logger.info(f'Final request statistics: {final_statistics}')
+        self._logger.info('Final request statistics: ')
+        final_statistics.to_table()
 
         return final_statistics