+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
-## 0.1.1 (Unreleased)
+## 0.1.2 (Unreleased)
+
+...
+
+## [0.1.1](https://github.com/apify/crawlee-python/releases/tag/v0.1.1) (2024-07-19)
+
+### Features
+
+- Support for proxy configuration in `PlaywrightCrawler`.
+- Blocking detection in `PlaywrightCrawler`.
+- Expose `crawler.log` to public.
 
 ### Bug fixes
 
 - Fix Pylance `reportPrivateImportUsage` errors by defining `__all__` in modules `__init__.py`.
 - Set `HTTPX` logging level to `WARNING` by default.
+- Fix CLI behavior with existing project folders.
 
 ## [0.1.0](https://github.com/apify/crawlee-python/releases/tag/v0.1.0) (2024-07-09)